requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.6.0
plotly>=5.20.0
pandas>=2.0.0
numpy>=1.26.0
google-generativeai>=0.7.0
torch>=2.0.0
transformers>=4.40.0
python-dotenv>=1.0.0
